% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{add_dup_markers}
\alias{add_dup_markers}
\title{Add duplicate markers to a map}
\usage{
add_dup_markers(maplist, bin_list)
}
\arguments{
\item{maplist}{A list of maps. Output of MDSMap_from_list.}

\item{bin_list}{A list of marker bins containing marker duplicates. Output of screen_for_duplicate_markers.}
}
\value{
A maplist with added markers.
}
\description{
Add duplicate markers to a map
}
