% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_datasets.R
\docType{data}
\name{ALL_dosages}
\alias{ALL_dosages}
\alias{segregating_data}
\alias{screened_data}
\alias{screened_data2}
\alias{screened_data3}
\alias{TRI_dosages}
\title{A dosage matrix for a random pairing tetraploid with five linkage groups.}
\format{
A matrix

An object of class \code{matrix} (inherits from \code{array}) with 2873 rows and 209 columns.

An object of class \code{matrix} (inherits from \code{array}) with 1417 rows and 209 columns.

An object of class \code{matrix} (inherits from \code{array}) with 1417 rows and 207 columns.

An object of class \code{matrix} (inherits from \code{array}) with 1417 rows and 200 columns.

An object of class \code{matrix} (inherits from \code{array}) with 250 rows and 202 columns.
}
\usage{
ALL_dosages

segregating_data

screened_data

screened_data2

screened_data3

TRI_dosages
}
\description{
A dosage matrix for a random pairing tetraploid with five linkage groups.
}
\keyword{datasets}
