% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_datasets.R
\docType{data}
\name{all_linkages_list_P1}
\alias{all_linkages_list_P1}
\alias{all_linkages_list_P1_split}
\alias{all_linkages_list_P1_subset}
\title{A (nested) list of linkage data frames classified per linkage group and homologue}
\format{An object of class \code{list} of length 5.}
\usage{
all_linkages_list_P1

all_linkages_list_P1_split

all_linkages_list_P1_subset
}
\description{
A (nested) list of linkage data frames classified per linkage group and homologue
}
\keyword{datasets}
