% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{calc_binning_thresholds}
\alias{calc_binning_thresholds}
\title{calc_binning_thresholds}
\usage{
calc_binning_thresholds(dosage_matrix)
}
\arguments{
\item{dosage_matrix}{An integer matrix with markers in rows and individuals in columns.}
}
\description{
Calculate LOD and recombination frequency thresholds for marker binning
}
