% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{calc_q1}
\alias{calc_q1}
\title{calc_q1 ***********************************************************}
\usage{
calc_q1(Pvalue, fracInvalid, bestfit, bestParentfit, Pvalue_threshold,
  fracInvalid_threshold)
}
\arguments{
\item{Pvalue}{P-value of bestParentfit segtype}

\item{fracInvalid}{Fraction invalid scores of bestParentfit segtype}

\item{bestfit}{segtype of best fit of population}

\item{bestParentfit}{segtype of best fit of parents}

\item{Pvalue_threshold}{Threshold P-value of bestParentfit segtype}

\item{fracInvalid_threshold}{a maximum threshold for the fracInvalid of the bestParentfit segtype
(with a larger fraction of invalid dosages in the F1 the q1 quality parameter will be set to 0)}
}
\description{
Calculate q1 quality score. q1 is a measure of the fit of selfit. 
It is based on Pvalue and the fraction of valid F1 scores. If either is below their 
threshold the q1 value is 0. Also, if bestParentfit is different from bestfit this indicates a segregation distortion.
}
