% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{calc_q2}
\alias{calc_q2}
\title{calc_q2 ***********************************************************}
\usage{
calc_q2(par.conflicts, par.NAfrac, matchParents, parentsScoredWithF1)
}
\arguments{
\item{par.conflicts}{Amount of conflicts in parental scores}

\item{par.NAfrac}{Amount of missing values in the parents}

\item{matchParents}{Should parents match F1? One of the following: "No", "Unknown", "OneOK" or "Yes"}

\item{parentsScoredWithF1}{Logical, if \code{TRUE} then q2 is determined not by matchParents but 
by the amount of conflicts and missing values in the parents. If \code{FALSE}, par.conflicts and 
par.NAfrac are not used and the information on the match between parents and F1 segtype has to come 
from matchParents.}
}
\description{
q2 is about how well the parents are scored and fit with the selfit segtype
}
