% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{calc_q3}
\alias{calc_q3}
\title{calc_q3 ***********************************************************}
\usage{
calc_q3(F1.NAfrac, fracNA_threshold)
}
\arguments{
\item{F1.NAfrac}{Fraction of missing values in the F1}

\item{fracNA_threshold}{Threshold for fraction missing values}
}
\description{
#q3 is about the fraction missing F1 scores. If that is larger than NA.threshold, q3=0;
else q3 depends on the fraction scored F1's
}
