% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{calc_qall}
\alias{calc_qall}
\title{calc_qall ***********************************************************}
\usage{
calc_qall(Pvalue_threshold, fracInvalid_threshold, fracNA_threshold, Pvalue,
  fracInvalid, F1.NAfrac, matchParents, bestfit, bestParentfit, par.conflicts,
  par.NAfrac, critweight, parentsScoredWithF1)
}
\arguments{
\item{Pvalue_threshold}{Threshold can be disabled by setting to 0}

\item{fracInvalid_threshold}{Threshold can be disabled by setting to 1}

\item{fracNA_threshold}{Threshold can be disabled by setting to 1}

\item{Pvalue}{P-value of bestParentfit segtype}

\item{fracInvalid}{Fraction invalid scores of the bestParentfit segtype}

\item{F1.NAfrac}{fraction missing scores among F1 samples}

\item{matchParents}{Should parents match F1? One of the following: "No", "Unknown", "OneOK" or "Yes"}

\item{bestfit}{segtype of best fit of population}

\item{bestParentfit}{segtype of best fit of parents}

\item{par.conflicts}{Logical vector: is there a conflict for P1, P2}

\item{par.NAfrac}{number of NA scores for the samples of P1, P2}

\item{critweight}{numeric vector of length 3 (3 quality criteria) with their weights, 
or NA (in that case qall = q1*q2*q3, not a weighted average)}

\item{parentsScoredWithF1}{Were parents scored with F1?}
}
\value{
vector of length 4 or 5. The last (two) components of the vector are qall_mult 
(obtained by multiplication of q1..q3), and if critweights is supplied, qall_weights 
(a weighted average of q1..q3)
}
\description{
calc_qall ***********************************************************
}
