% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{checkFilename}
\alias{checkFilename}
\title{checkFilename ***********************************************************}
\usage{
checkFilename(filename, overwrite = TRUE)
}
\arguments{
\item{filename}{The filename to check}

\item{overwrite}{if TRUE (default) an existing file of that name will be deleted (if it is not
protected by being locked, read-only, owned by another user etc)}
}
\description{
Check if a filename is valid for writing by trying to create the file.
If the filename contains a path, result will only be TRUE if the whole path already exists
}
