% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_datasets.R
\docType{data}
\name{chk1}
\alias{chk1}
\title{Example output of the checkF1 function}
\format{
An object of class \code{list} of length 2.
}
\usage{
chk1
}
\description{
Example output of the checkF1 function
}
\keyword{datasets}
