% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{chk2integer}
\alias{chk2integer}
\title{chk2integer ***********************************************************}
\usage{
chk2integer(chk)
}
\arguments{
\item{chk}{a data frame as returned by checkF1}
}
\value{
the same data frame, but with all columns from Parent1 to last ancestor 
converted to integer (as these are sometimes factors with levels not equal to values 
after reading checkF1 from file)
}
\description{
Convert data from checkF1 to integer
}
