% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{colour.bar}
\alias{colour.bar}
\title{Add colour bar scale to heatplot}
\usage{
colour.bar(col.data, min, max = -min, cex.ticks = 1.2, nticks = 11,
  ticks = seq(min, max, len = nticks), title = "", ylab = "",
  cex.lab = 1)
}
\arguments{
\item{col.data}{vector of colours}

\item{min}{minimum colour}

\item{max}{maximum colour}

\item{cex.ticks}{size of ticks on colour bar}

\item{nticks}{number of ticks on colour bar}

\item{ticks}{vector of positions of ticks on colour bar}

\item{title}{optional title for colour bar}

\item{ylab}{optional y-axis label for colour bar}

\item{cex.lab}{size of labels on colour bar}
}
\description{
Function to generate a scale for heatplots
}
