% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{convert_updog}
\alias{convert_updog}
\title{Convert (probabilistic) genotype calling results from updog to input compatible with polymapR.}
\usage{
convert_updog(mout)
}
\arguments{
\item{mout}{An object of class multidog; output of the function \link[updog]{multidog}.}
}
\value{
A data frame which include columns: 
MarkerName, SampleName,P0 ~ Pploidy (e.g. P0 ~ P4 for tetraploid, which represents
the probability assigning to this dosage), maxgeno (the most likely dosage),
and maxP (the maximum probability)
}
\description{
Convert (probabilistic) genotype calling results from updog to input compatible with polymapR.
}
\examples{
data("mout")
convert_updog(mout)
}
