% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{createChmHomList}
\alias{createChmHomList}
\title{User interface for specifying linkage group for a homologue}
\usage{
createChmHomList(LG_hom_stack, LG_number = 12)
}
\arguments{
\item{LG_hom_stack}{A data.frame defining clusters}

\item{LG_number}{Number of chromosomes (linkage groups)}
}
\description{
User interface for specifying linkage group for a homologue
}
