% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{digamfrq}
\alias{digamfrq}
\title{digamfrq ***********************************************************}
\usage{
digamfrq(ploidy, dosage)
}
\arguments{
\item{ploidy}{The ploidy level of the parent}

\item{dosage}{The parental dosage}
}
\description{
Get the integer ratios of gametes with dosage 0:(ploidy/2)
from an allopolyploid parent (disomic inheritance) with given ploidy and dosage
}
