% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_datasets.R
\docType{data}
\name{exampleRAD_mapping}
\alias{exampleRAD_mapping}
\title{Example output dataset of polyRAD::PipelineMapping2Parents function}
\format{
An object of class \code{RADdata} of length 23.
}
\usage{
exampleRAD_mapping
}
\description{
Example output dataset of polyRAD::PipelineMapping2Parents function
}
\keyword{datasets}
