% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{gcd}
\alias{gcd}
\title{gcd ***********************************************************}
\usage{
gcd(x, y)
}
\arguments{
\item{x}{integer (or vector of integers)}

\item{y}{integer (or vector of integers)}
}
\description{
Return the greatest common divisor of two integers (vectorized: x and y may be vectors)
}
