% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{gcd_all}
\alias{gcd_all}
\title{gcd_all ***********************************************************}
\usage{
gcd_all(x)
}
\arguments{
\item{x}{Input vector}
}
\description{
Return the greatest common divisor of all elements of vector x
}
