% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{get_markertype_combinations}
\alias{get_markertype_combinations}
\title{Visualize and get all markertype combinations for which there are functions in polymapR}
\usage{
get_markertype_combinations(ploidy, pairing,
  nonavailable_combinations = TRUE)
}
\arguments{
\item{ploidy}{Ploidy level}

\item{pairing}{Type of pairing. Either "random" or "preferential".}

\item{nonavailable_combinations}{Logical. Should nonavailable combinations be plotted with grey lines?}
}
\value{
A matrix with two columns. Each row represents a function with the first and second markertype.
}
\description{
Visualize and get all markertype combinations for which there are functions in polymapR
}
\examples{
get_markertype_combinations(4, "random")
}
