% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{getMatchParents}
\alias{getMatchParents}
\title{getMatchParents ***********************************************************}
\usage{
getMatchParents(parGeno, seginfoItem)
}
\arguments{
\item{parGeno}{integer vector of length 2 with the two parental (consensus) dosages (can each be NA)}

\item{seginfoItem}{one of the items (segtypes) of a list as returned by \code{\link{calcSegtypeInfo}}}
}
\description{
Does the specified segtype match the parental genotypes?
}
