% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_datasets.R
\docType{data}
\name{gp_df}
\alias{gp_df}
\title{An example of a genotype probability data frame}
\format{
Data frame
}
\usage{
gp_df
}
\description{
An example of a genotype probability data frame
}
\keyword{datasets}
