% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_datasets.R
\docType{data}
\name{gp_vignette_data}
\alias{gp_vignette_data}
\title{A list of objects needed to build the probabilistic genotype vignette}
\format{
An object of class \code{list} of length 15.
}
\usage{
gp_vignette_data
}
\description{
A list of objects needed to build the probabilistic genotype vignette
}
\keyword{datasets}
