% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_datasets.R
\docType{data}
\name{integrated.maplist}
\alias{integrated.maplist}
\title{A nested list with integrated maps}
\format{
An object of class \code{list} of length 5.
}
\usage{
integrated.maplist
}
\description{
A nested list with integrated maps
}
\keyword{datasets}
