% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{leftstr}
\alias{leftstr}
\title{Get substrings from the lefthand side}
\usage{
leftstr(x, n)
}
\arguments{
\item{x}{a character vector (or something having an as.character method)}

\item{n}{a single  number: if n>=0, the leftmost n characters of each element
of x are selected, if n<0 the (-n) rightmost characters are omitted}
}
\value{
character vector with leftside substrings of x
}
\description{
Get substrings from the lefthand side
}
