% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_datasets.R
\docType{data}
\name{LGHomDf_P1_1}
\alias{LGHomDf_P1_1}
\alias{LGHomDf_P2_1}
\alias{LGHomDf_P2_2}
\title{A \code{data.frame} specifying the assigned homologue and linkage group number per SxN marker}
\format{
\itemize{
 \item SxN_Marker. Markername of simplex nulliplex marker
 \item homologue. Assigned homologue number
 \item LG Assigned. linkage group number
}

An object of class \code{data.frame} with 195 rows and 3 columns.

An object of class \code{data.frame} with 195 rows and 3 columns.
}
\usage{
LGHomDf_P1_1

LGHomDf_P2_1

LGHomDf_P2_2
}
\description{
A \code{data.frame} specifying the assigned homologue and linkage group number per SxN marker
}
\keyword{datasets}
