% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{linterpol}
\alias{linterpol}
\title{linterpol ***********************************************************}
\usage{
linterpol(x, pnt1, pnt2)
}
\arguments{
\item{x}{x-coordinate for which corresponding y-coordinate is sought.}

\item{pnt1}{Point 1}

\item{pnt2}{Point 2}
}
\description{
linear interpolation: returns the y values matching the x values (vector)
on the line through points pnt1 and pnt2 (both are vectors of length 2)
possible error if both X-coordinates equal not checked or caught
}
