% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{makeProgeny}
\alias{makeProgeny}
\title{makeProgeny ***********************************************************}
\usage{
makeProgeny(gametes1, gametes2)
}
\arguments{
\item{gametes1}{vector of the integer ratios of the gametes produced by parent 1}

\item{gametes2}{vector of the integer ratios of the gametes produced by parent 2}
}
\description{
Returns a vector with the integer ratios of the F1 progeny
}
