% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_datasets.R
\docType{data}
\name{map2}
\alias{map2}
\title{A sample map}
\format{An object of class \code{data.frame} with 100 rows and 2 columns.}
\usage{
map2
}
\description{
A sample map
}
\keyword{datasets}
