% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_datasets.R
\docType{data}
\name{map3}
\alias{map3}
\title{A sample map}
\format{
An object of class \code{data.frame} with 60 rows and 2 columns.
}
\usage{
map3
}
\description{
A sample map
}
\keyword{datasets}
