% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_datasets.R
\docType{data}
\name{maplist_P1}
\alias{maplist_P1}
\alias{maplist_P1_subset}
\alias{maplist_P2_subset}
\title{A list of maps of one parent}
\format{An object of class \code{list} of length 5.}
\usage{
maplist_P1

maplist_P1_subset

maplist_P2_subset
}
\description{
A list of maps of one parent
}
\keyword{datasets}
