% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{Mode}
\alias{Mode}
\title{Calculate the mode of a vector}
\usage{
Mode(x)
}
\arguments{
\item{x}{Vector input, can be numeric, character or factor}
}
\description{
Calculate the mode of a vector
}
