% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_datasets.R
\docType{data}
\name{mout}
\alias{mout}
\title{Example output dataset of updog::multidog function}
\format{
An object of class \code{multidog} of length 2.
}
\usage{
mout
}
\description{
Example output dataset of updog::multidog function
}
\keyword{datasets}
