% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_datasets.R
\docType{data}
\name{P1_homologues}
\alias{P1_homologues}
\alias{P2_homologues}
\alias{P2_homologues_triploid}
\title{A list of cluster stacks at different LOD scores}
\format{
A list with with LOD thresholds as names. The list contains dataframes with the following format:
\itemize{
 \item marker. markername
 \item pseudohomologue. name of (pseudo)homologue
 }

An object of class \code{list} of length 10.

An object of class \code{list} of length 15.
}
\usage{
P1_homologues

P2_homologues

P2_homologues_triploid
}
\description{
A list of cluster stacks at different LOD scores
}
\keyword{datasets}
