% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_datasets.R
\docType{data}
\name{phased.maplist}
\alias{phased.maplist}
\title{A list of phased maps}
\format{An object of class \code{list} of length 5.}
\usage{
phased.maplist
}
\description{
A list of phased maps
}
\keyword{datasets}
