% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{plot_hom_vs_LG}
\alias{plot_hom_vs_LG}
\title{Plot homologue position versus integrated positions}
\usage{
plot_hom_vs_LG(map_df, maplist_homologue)
}
\arguments{
\item{map_df}{A dataframe of a map that defines a linkage group.}

\item{maplist_homologue}{A list of maps were each item represents a homoloogue.}
}
\description{
Plot homologue position versus integrated positions
}
\examples{
data("integrated.maplist", "maplist_P1_subset")
colnames(integrated.maplist[["LG2"]]) <- c("marker", "position", "QTL_LOD")
plot_hom_vs_LG(map_df = integrated.maplist[["LG2"]],
               maplist_homologue = maplist_P1_subset[["LG2"]])
}
