% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{plot_SNlinks}
\alias{plot_SNlinks}
\title{Plot links between 1.0 markers}
\usage{
plot_SNlinks(linkage_df, LG_hom_stack, LG_number, h1, h2, ymax = NULL)
}
\arguments{
\item{linkage_df}{A linkage data.frame.}

\item{LG_hom_stack}{A \code{data.frame} as a result of \code{\link{bridgeHomologues}}}

\item{LG_number}{Chromosome (LG) number.}

\item{h1}{Homologue to be compared.}

\item{h2}{Homologue to be compared against h1}

\item{ymax}{Maximum limit of y-axis}
}
\description{
Make a plot with recombination frequency versus LOD score.
}
