% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{polygamfrq}
\alias{polygamfrq}
\title{polygamfrq ***********************************************************}
\usage{
polygamfrq(ploidy, dosage)
}
\arguments{
\item{ploidy}{The ploidy level of the parent}

\item{dosage}{The parental dosage}
}
\description{
Get the integer ratios of gametes with dosage 0:(ploidy/2)
from an autopolyploid parent (polysomic inheritance) with given ploidy and dosage
the result can be converted to fractions by y <- x/(sum(x))
}
