% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polymapR-package.R
\docType{package}
\name{polymapR-package}
\alias{polymapR}
\alias{polymapR-package}
\title{polymapR: Linkage Analysis in Outcrossing Polyploids}
\description{
Creation of linkage maps in polyploid species from marker dosage scores of an F1 cross from two heterozygous parents. Currently works for outcrossing diploid, autotriploid, autotetraploid and autohexaploid species, as well as segmental allotetraploids. Methods are described in a manuscript of Bourke et al. (2018) \doi{10.1093/bioinformatics/bty371}. Since version 1.1.0, both discrete and probabilistic genotypes are acceptable input; for more details on the latter see Liao et al. (2021) \doi{10.1007/s00122-021-03834-x}.
}
\author{
\strong{Maintainer}: Peter Bourke \email{pbourkey@gmail.com}

Authors:
\itemize{
  \item Geert van Geest
}

Other contributors:
\itemize{
  \item Roeland Voorrips [contributor]
  \item Yanlin Liao [contributor]
}

}
\keyword{internal}
