% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\docType{package}
\name{polymapR}
\alias{polymapR}
\alias{polymapR-package}
\title{Linkage analysis in polyploids}
\description{
This package uses dosage-scored SNP markers from an F1 cross to perform linkage analysis in polyploids
}
