% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{prepare_pwd}
\alias{prepare_pwd}
\title{Prepare pwd for mapping scripts e.g. MDSmap}
\usage{
prepare_pwd(pwd)
}
\arguments{
\item{pwd}{A pwd data.frame}
}
\description{
Prepare a dataframe with pairwise recombination frequency and LOD score for mapping by re-ordering
}
