% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_functions.R
\name{r6_functions}
\alias{r6_functions}
\alias{r6_1.0_1.0}
\alias{r6_1.0_1.1}
\alias{r6_1.0_1.2}
\alias{r6_1.0_1.3}
\alias{r6_1.0_1.4}
\alias{r6_1.0_1.5}
\alias{r6_1.0_2.0}
\alias{r6_1.0_2.1}
\alias{r6_1.0_2.2}
\alias{r6_1.0_2.3}
\alias{r6_1.0_2.4}
\alias{r6_1.0_3.0}
\alias{r6_1.0_3.1}
\alias{r6_1.0_3.2}
\alias{r6_1.0_3.3}
\alias{r6_1.0_4.1}
\alias{r6_1.1_1.1}
\alias{r6_1.1_1.2}
\alias{r6_1.1_1.3}
\alias{r6_1.1_1.4}
\alias{r6_1.1_1.5}
\alias{r6_1.1_2.0}
\alias{r6_1.1_2.2}
\alias{r6_1.1_2.3}
\alias{r6_1.1_2.4}
\alias{r6_1.1_3.0}
\alias{r6_1.1_3.3}
\alias{r6_1.2_1.2}
\alias{r6_1.2_1.3}
\alias{r6_1.2_1.4}
\alias{r6_1.2_1.5}
\alias{r6_1.2_2.1}
\alias{r6_1.2_2.2}
\alias{r6_1.2_2.3}
\alias{r6_1.2_2.4}
\alias{r6_1.2_3.0}
\alias{r6_1.2_3.1}
\alias{r6_1.2_3.2}
\alias{r6_1.2_3.3}
\alias{r6_1.2_4.1}
\alias{r6_1.3_1.3}
\alias{r6_1.3_1.4}
\alias{r6_1.3_1.5}
\alias{r6_1.3_2.2}
\alias{r6_1.3_2.3}
\alias{r6_1.3_2.4}
\alias{r6_1.3_3.1}
\alias{r6_1.3_3.2}
\alias{r6_1.3_3.3}
\alias{r6_1.3_4.1}
\alias{r6_1.4_1.4}
\alias{r6_1.4_1.5}
\alias{r6_1.4_2.2}
\alias{r6_1.4_2.3}
\alias{r6_1.4_2.4}
\alias{r6_1.4_3.2}
\alias{r6_1.4_3.3}
\alias{r6_1.4_4.1}
\alias{r6_1.5_1.5}
\alias{r6_1.5_2.2}
\alias{r6_1.5_2.3}
\alias{r6_1.5_2.4}
\alias{r6_1.5_3.2}
\alias{r6_1.5_3.3}
\alias{r6_2.0_1.2}
\alias{r6_2.0_1.3}
\alias{r6_2.0_1.4}
\alias{r6_2.0_1.5}
\alias{r6_2.0_2.0}
\alias{r6_2.0_2.1}
\alias{r6_2.0_2.2}
\alias{r6_2.0_2.3}
\alias{r6_2.0_2.4}
\alias{r6_2.0_3.0}
\alias{r6_2.0_3.1}
\alias{r6_2.0_3.2}
\alias{r6_2.0_3.3}
\alias{r6_2.0_4.1}
\alias{r6_2.1_1.5}
\alias{r6_2.1_2.4}
\alias{r6_2.1_3.0}
\alias{r6_2.2_2.2}
\alias{r6_2.2_2.3}
\alias{r6_2.2_2.4}
\alias{r6_2.2_3.2}
\alias{r6_2.2_3.3}
\alias{r6_2.3_2.3}
\alias{r6_2.3_2.4}
\alias{r6_2.3_3.2}
\alias{r6_2.3_3.3}
\alias{r6_2.4_2.4}
\alias{r6_3.0_1.3}
\alias{r6_3.0_1.4}
\alias{r6_3.0_1.5}
\alias{r6_3.0_2.2}
\alias{r6_3.0_2.3}
\alias{r6_3.0_2.4}
\alias{r6_3.0_3.0}
\alias{r6_3.0_3.1}
\alias{r6_3.0_3.2}
\alias{r6_3.0_3.3}
\alias{r6_3.0_4.1}
\alias{r6_3.3_2.4}
\alias{r6_3.3_3.3}
\alias{p6_1.0_1.0}
\title{Calculate recombination frequency, LOD and log-likelihood from frequency tables in a random pairing hexaploid}
\usage{
r6_1.0_1.0(x)

r6_1.0_1.1(x)

r6_1.0_1.2(x)

r6_1.0_1.3(x)

r6_1.0_1.4(x)

r6_1.0_1.5(x)

r6_1.0_2.0(x)

r6_1.0_2.1(x)

r6_1.0_2.2(x)

r6_1.0_2.3(x)

r6_1.0_2.4(x)

r6_1.0_3.0(x)

r6_1.0_3.1(x)

r6_1.0_3.2(x)

r6_1.0_3.3(x)

r6_1.0_4.1(x)

r6_1.1_1.1(x)

r6_1.1_1.2(x)

r6_1.1_1.3(x)

r6_1.1_1.4(x)

r6_1.1_1.5(x)

r6_1.1_2.0(x)

r6_1.1_2.2(x)

r6_1.1_2.3(x)

r6_1.1_2.4(x)

r6_1.1_3.0(x)

r6_1.1_3.3(x)

r6_1.2_1.2(x)

r6_1.2_1.3(x)

r6_1.2_1.4(x)

r6_1.2_1.5(x)

r6_1.2_2.1(x)

r6_1.2_2.2(x)

r6_1.2_2.3(x)

r6_1.2_2.4(x)

r6_1.2_3.0(x)

r6_1.2_3.1(x)

r6_1.2_3.2(x)

r6_1.2_3.3(x)

r6_1.2_4.1(x)

r6_1.3_1.3(x)

r6_1.3_1.4(x)

r6_1.3_1.5(x)

r6_1.3_2.2(x)

r6_1.3_2.3(x)

r6_1.3_2.4(x)

r6_1.3_3.1(x)

r6_1.3_3.2(x)

r6_1.3_3.3(x)

r6_1.3_4.1(x)

r6_1.4_1.4(x)

r6_1.4_1.5(x)

r6_1.4_2.2(x)

r6_1.4_2.3(x)

r6_1.4_2.4(x)

r6_1.4_3.2(x)

r6_1.4_3.3(x)

r6_1.4_4.1(x)

r6_1.5_1.5(x)

r6_1.5_2.2(x)

r6_1.5_2.3(x)

r6_1.5_2.4(x)

r6_1.5_3.2(x)

r6_1.5_3.3(x)

r6_2.0_1.2(x)

r6_2.0_1.3(x)

r6_2.0_1.4(x)

r6_2.0_1.5(x)

r6_2.0_2.0(x)

r6_2.0_2.1(x)

r6_2.0_2.2(x)

r6_2.0_2.3(x)

r6_2.0_2.4(x)

r6_2.0_3.0(x)

r6_2.0_3.1(x)

r6_2.0_3.2(x)

r6_2.0_3.3(x)

r6_2.0_4.1(x)

r6_2.1_1.5(x)

r6_2.1_2.4(x)

r6_2.1_3.0(x)

r6_2.2_2.2(x)

r6_2.2_2.3(x)

r6_2.2_2.4(x)

r6_2.2_3.2(x)

r6_2.2_3.3(x)

r6_2.3_2.3(x)

r6_2.3_2.4(x)

r6_2.3_3.2(x)

r6_2.3_3.3(x)

r6_2.4_2.4(x)

r6_3.0_1.3(x)

r6_3.0_1.4(x)

r6_3.0_1.5(x)

r6_3.0_2.2(x)

r6_3.0_2.3(x)

r6_3.0_2.4(x)

r6_3.0_3.0(x)

r6_3.0_3.1(x)

r6_3.0_3.2(x)

r6_3.0_3.3(x)

r6_3.0_4.1(x)

r6_3.3_2.4(x)

r6_3.3_3.3(x)

p6_1.0_1.0(x)
}
\arguments{
\item{x}{A frequency table of the different classes of dosages in the progeny. The column names start with \code{"n_"}. Followed by the dosage of the first marker and then of the second.}
}
\value{
A list with the following items:
\item{r_mat}{A matrix with recombination frequencies for the different phases}
\item{LOD_mat}{A matrix with LOD scores for the different phases}
\item{logL_mat}{A matrix with log likelihood ratios for the different phases}
\item{phasing_strategy}{A character string specifying the phasing strategy. \code{"MLL"} for maximum likelihood en \code{"MINR"} for minimum recombination frequency.}
\item{possible_phases}{The phases between markers that are possible. Same order and length as column names of output matrices.}
}
\description{
This group of functions is called by \code{\link{linkage}}.
}
