% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{rightstr}
\alias{rightstr}
\title{Get substrings from the righthand side}
\usage{
rightstr(x, n)
}
\arguments{
\item{x}{a character vector (or something having an as.character method)}

\item{n}{a single  number: if n>=0, the rightmost n characters of each element
of x are selected, if n<0 the (-n) leftmost characters are omitted}
}
\value{
character vector with rightside substrings of x
}
\description{
Get substrings from the righthand side
}
