% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_datasets.R
\docType{data}
\name{SN_SN_P1}
\alias{SN_SN_P1}
\alias{SN_SN_P2}
\alias{SN_SS_P1}
\alias{SN_SS_P2}
\alias{SN_DN_P1}
\alias{SN_DN_P2}
\alias{SN_SN_P2_triploid}
\title{A linkage \code{data.frame}.}
\format{
\itemize{
  \item marker_a. First marker in comparison
  \item marker_b. Second marker in comparison
  \item r. recombination frequency
  \item LOD. LOD score
  \item phase. The phase between markers
}

An object of class \code{linkage_df} (inherits from \code{data.frame}) with 19306 rows and 5 columns.

An object of class \code{linkage_df} (inherits from \code{data.frame}) with 53152 rows and 5 columns.

An object of class \code{linkage_df} (inherits from \code{data.frame}) with 59494 rows and 5 columns.

An object of class \code{linkage_df} (inherits from \code{data.frame}) with 19536 rows and 5 columns.

An object of class \code{linkage_df} (inherits from \code{data.frame}) with 19897 rows and 5 columns.

An object of class \code{data.frame} with 6655 rows and 5 columns.
}
\usage{
SN_SN_P1

SN_SN_P2

SN_SS_P1

SN_SS_P2

SN_DN_P1

SN_DN_P2

SN_SN_P2_triploid
}
\description{
A linkage \code{data.frame}.
}
\keyword{datasets}
