% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{test_cluster_stack}
\alias{test_cluster_stack}
\title{Error and warning handling for cluster_stack}
\usage{
test_cluster_stack(cluster_stack)
}
\arguments{
\item{cluster_stack}{A data.frame with a column "marker" specifying markernames, 
and a column "cluster" specifying marker cluster.}
}
\description{
Error and warning handling for cluster_stack
}
