% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{test_dosage_matrix}
\alias{test_dosage_matrix}
\title{Error and warning handling for dosage_matrix}
\usage{
test_dosage_matrix(dosage_matrix)
}
\arguments{
\item{dosage_matrix}{An integer matrix with markers in rows and individuals in columns.}
}
\description{
Error and warning handling for dosage_matrix
}
