% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{test_LG_hom_stack}
\alias{test_LG_hom_stack}
\title{Error and warning handling for LG_hom_stack}
\usage{
test_LG_hom_stack(LG_hom_stack)
}
\arguments{
\item{LG_hom_stack}{A data.frame with a column "SxN_Marker" specifying markernames, 
a column "homologue" specifying homologue cluster and "LG" specifying linkage group.}
}
\description{
Error and warning handling for LG_hom_stack
}
