% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{test_linkage_df}
\alias{test_linkage_df}
\title{Error and warning handling for linkage_df}
\usage{
test_linkage_df(linkage_df)
}
\arguments{
\item{linkage_df}{A linkage data.frame as output of \code{\link{linkage}} calculating linkage between 1.0 markers.}
}
\description{
Error and warning handling for linkage_df
}
