% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{vector.to.matrix}
\alias{vector.to.matrix}
\title{Large vector or in standardized matrix}
\usage{
vector.to.matrix(x, n.columns)
}
\arguments{
\item{x}{A vector}

\item{n.columns}{Integer, number of columns}
}
\description{
Turns a vector in a matrix with fixed number of columns
}
