% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{write.logheader}
\alias{write.logheader}
\title{Write a header for the log file}
\usage{
write.logheader(matc, log)
}
\arguments{
\item{matc}{A object of class \code{call}}

\item{log}{A character string specifying the log file}
}
\description{
Functionalized writing of function name and arguments as start for log paragraph.
}
