% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{r_LOD_plot}
\alias{r_LOD_plot}
\title{Plot r versus LOD}
\usage{
r_LOD_plot(linkage_df, plot_main = "", chm = NA, r_max = 0.5)
}
\arguments{
\item{linkage_df}{A linkage data.frame as output of \code{\link{linkage}}.}

\item{plot_main}{A character string specifying the main title}

\item{chm}{Integer specifying chromosome}

\item{r_max}{Maximum r value to plot}
}
\description{
\code{r_LOD_plot} plots r versus LOD, colour separated for different phases.
}
\examples{
data("SN_SN_P1")
r_LOD_plot(SN_SN_P1)
}
