% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_datasets.R
\docType{data}
\name{SN_SN_P1}
\alias{SN_SN_P1}
\alias{SN_SN_P2}
\alias{SN_SS_P1}
\alias{SN_SS_P2}
\alias{SN_DN_P1}
\alias{SN_DN_P2}
\alias{SN_SN_P2_triploid}
\title{A linkage \code{data.frame}.}
\format{\itemize{
  \item marker_a. First marker in comparison
  \item marker_b. Second marker in comparison
  \item r. recombination frequency
  \item LOD. LOD score
  \item phase. The phase between markers
}}
\usage{
SN_SN_P1

SN_SN_P2

SN_SS_P1

SN_SS_P2

SN_DN_P1

SN_DN_P2

SN_SN_P2_triploid
}
\description{
A linkage \code{data.frame}.
}
\keyword{datasets}
